requirements.txt
aws-cdk-lib>=2.100.0
constructs>=10.0.0
boto3>=1.34.0
pytest>=7.4.0
pytest-cov>=4.1.0
moto[sns,sqs,dynamodb]>=4.2.0
